Move to more consistent location, fix runtime errors
